+++ packages/fiori-annotation-api/src/index.ts
@@ -5,4 +5,4 @@ export * from './types/change.js';
 export * from './external-services.js';
 export type { ServiceArtifacts } from './types/adapter.js';
 export { findAnnotation, getAvtNodeFromPointer } from './avt/index.js';
-export { getXmlServiceArtifacts, CdsAnnotationProvider } from './annotation-provider.js';
+export { getXmlServiceArtifacts, CdsAnnotationProvider, uniformUrl } from './annotation-provider.js';