+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.9.44
+
+### Patch Changes
+
+- 52f6549: Add .cds annotations support to enable linting of CAP apps with the eslint-plugin-fiori-tools.
+
 ## 0.9.43
 
 ### Patch Changes